simplify package structure
