+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/mutation/SimpleMutation.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.graphql.sample.mutation
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.spring.server.annotation.Mutation
+import com.expediagroup.graphql.spring.annotation.Mutation
 import org.springframework.stereotype.Component
 
 /**
