+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/mutation/WidgetMutation.kt
@@ -18,7 +18,7 @@ package com.expediagroup.graphql.sample.mutation
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.sample.model.Widget
-import com.expediagroup.graphql.spring.server.annotation.Mutation
+import com.expediagroup.graphql.spring.annotation.Mutation
 import org.springframework.stereotype.Component
 
 /**
