+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/query/AsyncQuery.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.graphql.sample.query
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.spring.server.annotation.Query
+import com.expediagroup.graphql.spring.annotation.Query
 import org.springframework.stereotype.Component
 import java.util.concurrent.CompletableFuture
