+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/query/CustomDirectiveQuery.kt
@@ -20,7 +20,7 @@ import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.sample.directives.LowercaseDirective
 import com.expediagroup.graphql.sample.directives.SpecificValueOnly
 import com.expediagroup.graphql.sample.directives.StringEval
-import com.expediagroup.graphql.spring.server.annotation.Query
+import com.expediagroup.graphql.spring.annotation.Query
 import org.springframework.stereotype.Component
 
 @Component
