+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/query/EnvironmentQuery.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.graphql.sample.query
 
-import com.expediagroup.graphql.spring.server.annotation.Query
+import com.expediagroup.graphql.spring.annotation.Query
 import graphql.schema.DataFetchingEnvironment
 import org.springframework.stereotype.Component
