+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/sample/query/SimpleQuery.kt
@@ -18,7 +18,7 @@ package com.expediagroup.graphql.sample.query
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.annotations.GraphQLIgnore
-import com.expediagroup.graphql.spring.server.annotation.Query
+import com.expediagroup.graphql.spring.annotation.Query
 import org.springframework.stereotype.Component
 import java.util.Random
