+++ graphql-kotlin-spring-server/src/main/kotlin/com/expediagroup/graphql/spring/exception/KotlinDataFetcherExceptionHandler.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.expediagroup.graphql.spring.server.exception
+package com.expediagroup.graphql.spring.exception
 
 import graphql.ExceptionWhileDataFetching
 import graphql.GraphQLError
