+++ graphql-kotlin-spring-server/src/main/resources/META-INF/spring-configuration-metadata.json
@@ -4,23 +4,23 @@
       "name": "graphql.endpoint",
       "type": "java.lang.String",
       "defaultValue": "graphql",
-      "sourceType": "com.expediagroup.graphql.spring.server.FederationConfigurationProperties"
+      "sourceType": "com.expediagroup.graphql.spring.FederationConfigurationProperties"
     },
     {
       "name": "graphql.packages",
       "type": "java.util.List<java.lang.String>",
-      "sourceType": "com.expediagroup.graphql.spring.server.FederationConfigurationProperties"
+      "sourceType": "com.expediagroup.graphql.spring.FederationConfigurationProperties"
     },
     {
       "name": "graphql.federation.enabled",
       "type": "java.lang.Boolean",
-      "sourceType": "com.expediagroup.graphql.spring.server.FederationConfigurationProperties"
+      "sourceType": "com.expediagroup.graphql.spring.FederationConfigurationProperties"
     },
     {
       "name": "graphql.subscriptions.endpoint",
       "type": "java.lang.String",
       "defaultValue": "subscriptions",
-      "sourceType": "com.expediagroup.graphql.spring.server.FederationConfigurationProperties"
+      "sourceType": "com.expediagroup.graphql.spring.FederationConfigurationProperties"
     }
   ]
 }
